+++ Sources/LSPTestSupport/TestJSONRPCConnection.swift
@@ -10,8 +10,6 @@
 //
 //===----------------------------------------------------------------------===//
 
-import SKCore
-import SKSupport
 import LanguageServerProtocol
 import LanguageServerProtocolJSONRPC
 import class Foundation.Pipe
@@ -79,14 +77,12 @@ public struct TestLocalConnection {
   }
 }
 
-public final class TestClient: LanguageServerEndpoint {
-
+public final class TestClient: MessageHandler {
   /// The connection to the language client.
   public let server: Connection
 
   public init(server: Connection) {
     self.server = server
-    super.init()
   }
 
   public var replyQueue: DispatchQueue = DispatchQueue(label: "testclient-reply-queue")
@@ -123,11 +119,9 @@ public final class TestClient: LanguageServerEndpoint {
     appendOneShotRequestHandler(handler)
   }
 
-  override public func _registerBuiltinHandlers() {
+  public func handle<N>(_ params: N, from clientID: ObjectIdentifier) where N: NotificationType {
+    let notification = Notification(params, clientID: clientID)
 
-  }
-
-  override public func _handleUnknown<N>(_ notification: Notification<N>) {
     guard !oneShotNotificationHandlers.isEmpty else {
       if allowUnexpectedNotification { return }
       fatalError("unexpected notification \(notification)")
@@ -136,7 +130,11 @@ public final class TestClient: LanguageServerEndpoint {
     handler(notification)
   }
 
-  override public func _handleUnknown<R>(_ request: Request<R>) where R : RequestType {
+  public func handle<R: RequestType>(_ params: R, id: RequestID, from clientID: ObjectIdentifier, reply: @escaping (LSPResult<R.Response>) -> Void) {
+    let cancellationToken = CancellationToken()
+
+    let request = Request(params, id: id, clientID: clientID, cancellation: cancellationToken, reply: reply)
+
     guard !oneShotRequestHandlers.isEmpty else {
       fatalError("unexpected request \(request)")
     }
@@ -211,32 +209,42 @@ extension TestClient: Connection {
   }
 }
 
-public final class TestServer: LanguageServer {
+public final class TestServer: MessageHandler {
+  public let client: Connection
 
-  override public func _registerBuiltinHandlers() {
-    register { (req: Request<EchoRequest>) in
-      req.reply(req.params.string)
+  init(client: Connection) {
+    self.client = client
+  }
+
+  public func handle<N: NotificationType>(_ params: N, from clientID: ObjectIdentifier) {
+    let note = Notification(params, clientID: clientID)
+    if params is EchoNotification {
+      self.client.send(note.params)
+    } else {
+      fatalError("Unhandled notification")
     }
+  }
 
-    register { (req: Request<EchoError>) in
+  public func handle<R: RequestType>(_ params: R, id: RequestID, from clientID: ObjectIdentifier, reply: @escaping (LSPResult<R.Response >) -> Void) {
+    let cancellationToken = CancellationToken()
+
+    if let params = params as? EchoRequest {
+      let req = Request(params, id: id, clientID: clientID, cancellation: cancellationToken, reply: { result in
+        reply(result.map({ $0 as! R.Response }))
+      })
+      req.reply(req.params.string)
+    } else if let params = params as? EchoError {
+      let req = Request(params, id: id, clientID: clientID, cancellation: cancellationToken, reply: { result in
+        reply(result.map({ $0 as! R.Response }))
+      })
       if let code = req.params.code {
         req.reply(.failure(ResponseError(code: code, message: req.params.message!)))
       } else {
         req.reply(VoidResponse())
       }
+    } else {
+      fatalError("Unhandled request")
     }
-
-    register { [unowned self] (note:  Notification<EchoNotification>) in
-      self.client.send(note.params)
-    }
-  }
-
-  override public func _handleUnknown<R>(_ request: Request<R>) {
-    fatalError()
-  }
-
-  override public func _handleUnknown<N>(_ notification:  Notification<N>) {
-    fatalError()
   }
 }
