+++ Sources/LanguageServerProtocolJSONRPC/JSONRPCConnection.swift
@@ -10,10 +10,10 @@
 //
 //===----------------------------------------------------------------------===//
 
-import SKSupport
-import LanguageServerProtocol
 import Dispatch
 import Foundation
+import LanguageServerProtocol
+import LSPSupport
 
 /// A connection between a message handler (e.g. language server) in the same process as the connection object and a remote message handler (e.g. language client) that may run in another process using JSON RPC messages sent over a pair of in/out file descriptors.
 ///
