+++ Sources/SourceKit/sourcekitd/Diagnostic.swift
@@ -11,6 +11,7 @@
 //===----------------------------------------------------------------------===//
 
 import LanguageServerProtocol
+import LSPSupport
 import SKSupport
 import sourcekitd
