+++ Tests/SourceKitTests/DocumentSymbolTests.swift
@@ -11,10 +11,10 @@
 //===----------------------------------------------------------------------===//
 
 import LanguageServerProtocol
-import SKSupport
+import LSPTestSupport
 import SKTestSupport
-import XCTest
 import SourceKit
+import XCTest
 
 final class DocumentSymbolTest: XCTestCase {
   typealias DocumentSymbolCapabilities = TextDocumentClientCapabilities.DocumentSymbol
